requirements.txt
Flask==0.11.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-WTF==0.11
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.9.0
SQLAlchemy==1.0.7
WTForms==2.0.2
Werkzeug==0.10.4
alembic==0.7.6
argparse==1.2.1
bbcode==1.0.21
blinker==1.3
decorator==3.4.2
fixture==1.5.2
itsdangerous==0.24
python-dateutil==2.4.2
six==1.9.0
wsgiref==0.1.2
lxml==3.4.4
AuthorizeSauce==0.4.1
us==0.9.1
#Flask-Captcha==0.1.8